remove global handler
